requirements.txt
africastalking==1.2.9
asgiref==3.8.1
backports.zoneinfo==0.2.1
black==24.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.1
cryptography==44.0.1
defusedxml==0.8.0rc2
Django==4.2.19
django-js-asset==2.2.0
django-model-utils==5.0.0
django-mptt==0.14.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
drf-yasg==1.21.8
ecdsa==0.19.0
exceptiongroup==1.2.2
flake8==5.0.4
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jwcrypto==1.5.6
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.9.1
pycparser==2.22
pyflakes==2.5.0
PyJWT==2.9.0
pytest==8.3.4
pytest-cov==5.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schema==0.7.7
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
